+++ fp.ts
@@ -26,7 +26,7 @@ export const filter = curryIterFunction(transformers.filter);
 export const indexedPairs = transformers.indexedPairs;
 export const chunkify = curryIterFunction(transformers.chunkify);
 export const remember = transformers.rememeber;
-export const flat = curryIterFunction(transformers.flat)
+export const flat = curryIterFunction(transformers.flat);
 export const completeFlat = transformers.completeFlat;
 
 // Effectors